ssm/requirements.txt
# common
redis==3.0.1
celery==4.2.1
flower==0.9.2
psycopg2-binary==2.7.6.1

# django
django==2.1.5
django-celery-email==2.0.1
django-model-utils==3.1.2
django-debug-toolbar==1.11
django-taggit==0.23.0
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework_simplejwt==3.3
djangorestframework-filters==v1.0.0.dev0  #FIXME: doesn't support django==2.* and django-filter==2.*
drf-dynamic-fields==0.3.0
